:recycle: extract Mbti component from SignupForm
